// packages/backend/prisma/generated/zod/outputTypeSchemas/RefreshTokenUpsertArgsSchema.ts
import { z } from 'zod';
import type { Prisma } from '@prisma/client';
import { RefreshTokenIncludeSchema } from '../inputTypeSchemas/RefreshTokenIncludeSchema'
import { RefreshTokenWhereUniqueInputSchema } from '../inputTypeSchemas/RefreshTokenWhereUniqueInputSchema'
import { RefreshTokenCreateInputSchema } from '../inputTypeSchemas/RefreshTokenCreateInputSchema'
import { RefreshTokenUncheckedCreateInputSchema } from '../inputTypeSchemas/RefreshTokenUncheckedCreateInputSchema'
import { RefreshTokenUpdateInputSchema } from '../inputTypeSchemas/RefreshTokenUpdateInputSchema'
import { RefreshTokenUncheckedUpdateInputSchema } from '../inputTypeSchemas/RefreshTokenUncheckedUpdateInputSchema'
import { UserArgsSchema } from "../outputTypeSchemas/UserArgsSchema"
// Select schema needs to be in file to prevent circular imports
//------------------------------------------------------

export const RefreshTokenSelectSchema: z.ZodType<Prisma.RefreshTokenSelect> = z.object({
  id: z.boolean().optional(),
  userId: z.boolean().optional(),
  tokenHash: z.boolean().optional(),
  expiresAt: z.boolean().optional(),
  createdAt: z.boolean().optional(),
  revokedAt: z.boolean().optional(),
  user: z.union([z.boolean(),z.lazy(() => UserArgsSchema)]).optional(),
}).strict()

export const RefreshTokenUpsertArgsSchema: z.ZodType<Prisma.RefreshTokenUpsertArgs> = z.object({
  select: RefreshTokenSelectSchema.optional(),
  include: z.lazy(() => RefreshTokenIncludeSchema).optional(),
  where: RefreshTokenWhereUniqueInputSchema, 
  create: z.union([ RefreshTokenCreateInputSchema, RefreshTokenUncheckedCreateInputSchema ]),
  update: z.union([ RefreshTokenUpdateInputSchema, RefreshTokenUncheckedUpdateInputSchema ]),
}).strict();

export default RefreshTokenUpsertArgsSchema;
